Sergey Molchun HW lesson 35 review points fix2
